:art: [refactor] : Image, Report, Rank Converter class 분리
